Query order for Repair Xducers & email no final test
Repair portal will now send emails when an Xducer is not in the final test table. Will also send emails for unhandled exceptions and errors that cause the program to close. Also added a check to ensure that if parts replaced items are to be added that some exist... need to verify conditions with Dexter though.